+++ CHANGES.rst
@@ -1,6 +1,8 @@
 2.4.1dev (unreleased)
 =====================
 
+
+- Updated interact features to work with JupyterHub (e.g. TiKE) [#1349]
 - Exposed niters parameter in the PCA function of design matrix
 - Fixed the aperture parsing functions inside TPFs to be compliant with `numpy` v1.25 [#1360]
 - Made `LombScarglePeriodogram` compatible with Astropy v5.3 [#1342]